Improve configuration samples and update the TODO file
Configuration samples updated to make it easier for the end user to
create a functional configuration.

More hints have been added to the comments in some of the configuration
samples.

More default settings have been added to IKEv1 default.conf to enable
end users to easily configure connections using the unspecified address
IP_ANY for the peer. This includes enabling pre-shared key
authentication for the "wildcard" address by default, with clear
comments explainng how to disable this if desired for security reasons.
This is enabled by default because the wildcard pre-shared key
is needed for the iphone L2TP/IPsec client that cannot use certificate
authentication.

TODO file updated to reflect progress made and describe further work
that needs to be done.