Handle case when client asks for a dual stack configuration, with one
IPv4 address and one IPv6 address. Previously, iked would only
send the client one leased address, even if the client requested two
addresses, an IPv4 address and an IPv6 address.

With this patch, iked will send both an IPv4 address and an IPv6
address to the client when the client requests this and racoon2 is
configured with both an IPv4 address pool and an IPv6 address pool.

Also, with this patch iked will send traffic selectors for both address
families if racoon2.conf includes traffic selectors for both address
families that match the leased addresses.

With this patch it is possible to confiugre the windows 10 ikev2 client
to tunnel both IPv4 and IPv6 traffic over a single pair of child
security associations to racoon2 configured as an IKEv2 VPN server
using tunnel mode, using the generic tunnel interface (gif) on
NetBSD.