Import changes to support deleting only the matching policy
Currently if there are multiple security policies using the same
selector index, when one of the connections is terminated the policies
for all the connections will be deleted. This patch provides support
to only delete the policies for the connection that is being terminated.

New code is not activated since it does not work properly yet.