Prevent segfault when freeing memory, enable selective deleting of...
security policies that use the same selector when destroying a child_sa.

When freeing the rcpfk_msg structure, do not free the pointers to the
sockaddr structures if they are set to the addresses of the internal
buffers. Without this patch, the new functions to enable selective
deleting of security policies cause a segmentation fault.